improve generated instruction layout for most cases
By having most control flow constructs of the language use the negated version of their condition instead of the condition itself,
we can prevent (most of the time) the ugly code generation that caused two parts of the same construct (e.g. a for loop) to be distant from each other on the generated code.
While that didn't  have any impact on performance, it still made the generated code harder to read and reason about.